+++ service/controller/key/common.go
@@ -20,8 +20,6 @@ const (
 	AWSCNIDefaultMinimumIPTarget = "40"
 	AWSCNIDefaultWarmIPTarget    = "10"
 
-	AWSCloudControllerManager = "aws-cloud-controller-manager"
-
 	ELBInstanceStateInService = "InService"
 
 	DrainerResyncPeriod = time.Minute * 2