+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Revert applying external cloud controller manager as a static pod.
+
 ## [12.1.0] - 2022-07-18
 
 ### Added